windows_server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-socketio>=5.10.0
jinja2>=3.1.2
